tools/requirements.txt
Pillow	# For reading and writing `.png` files
pynput	# For listening for keyboard input

py360convert == 0.1.0	# For making equirectangular panoramas